refactor(example): rename SomeClass to example
