+++ example/src/example.js
